requirements.txt
streamlit
langchain
langchain-core
langchain-huggingface
python-dotenv
python-telegram-bot
langchain-groq
